requirement.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.4
numpy==1.23.3
pandas==1.5.0
pipreqs==0.4.11
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
yarg==0.1.9
